requirements.txt
#devops
black==22.3.0
click==8.1.3 
pytest==7.1.3
pylint==2.17.5
pathlib==1.0.1
seaborn
matplotlib
Pillow
